+++ Examples/Resources/xSQLServerMaxDop/1-SetMaxDopToOne.ps1
@@ -1,0 +1,30 @@
+<#
+.EXAMPLE
+    This example shows how to set max degree of parallelism server
+    configuration option with the value equal to 1.
+#>
+Configuration Example
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
+        $SysAdminAccount
+    )
+
+    Import-DscResource -ModuleName xSqlServer
+
+    node localhost
+    {
+        xSQLServerMaxDop Set_SQLServerMaxDop_ToOne
+        {
+            Ensure = 'Present'
+            DynamicAlloc = $false
+            MaxDop = 1
+            SQLServer = 'SQLServer'
+            SQLInstanceName = 'DSC'
+            PsDscRunAsCredential = $SysAdminAccount
+        }
+    }
+}